pyproject.toml
[project]
name = "amonusdetector"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "albumentations>=2.0.8",
    "cairosvg>=2.8.2",
    "lightning>=2.6.0",
    "numpy>=2.3.5",
    "opencv-python>=4.11.0.86",
    "pandas>=2.3.3",
    "pillow>=12.0.0",
    "pre-commit>=4.5.0",
    "progressbar2>=4.5.0",
    "pytubefix>=10.3.5",
    "tensorboard>=2.20.0",
    "torchvision>=0.24.1",
]
